% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendario.R
\name{Calendario}
\alias{Calendario}
\title{R6 Class Representing Tasks and Projects}
\description{
A calendario object tracks a collection of projects that are contain
tasks that are to be performed at different times
}
\details{
BLAH BLAH BLAH
}
\examples{

## ------------------------------------------------
## Method `Calendario$get_tasks`
## ------------------------------------------------

cal <- Calendario$new()
cal$set_project("my project")
cal$add_task("first task", "jan 21")
cal$add_task("second task", "jan 22")
cal$get_tasks()


## ------------------------------------------------
## Method `Calendario$set_project`
## ------------------------------------------------

cal <- Calendario$new()
cal$set_project("my project")
cal$get_project()


## ------------------------------------------------
## Method `Calendario$set_defaults`
## ------------------------------------------------

cal <- Calendario$new()
cal$set_defaults(type = "fun", team = "danielle")
cal$get_defaults()


## ------------------------------------------------
## Method `Calendario$get_project`
## ------------------------------------------------

cal <- Calendario$new()
cal$set_project("my project")
cal$get_project()

## ------------------------------------------------
## Method `Calendario$get_defaults`
## ------------------------------------------------

cal <- Calendario$new()
cal$get_defaults()

## ------------------------------------------------
## Method `Calendario$get_options`
## ------------------------------------------------

cal <- Calendario$new()
cal$get_options()

## ------------------------------------------------
## Method `Calendario$add_task`
## ------------------------------------------------

cal <- Calendario$new()
cal$add_task(
  description = "label for the task",
  start = as.Date("2025-01-24"),
  stop = as.Date("2025-01-26"),
  daily_hours = 1,
  project = "my project",
  team = "just me"
)
cal$get_tasks()

## ------------------------------------------------
## Method `Calendario$show_tasks`
## ------------------------------------------------

cal <- Calendario$new()
cal$set_project("my project")
cal$add_task("first task", "jan 21")
cal$add_task("second task", "jan 23")
cal$show_tasks()


## ------------------------------------------------
## Method `Calendario$get_workload`
## ------------------------------------------------

cal <- Calendario$new()
cal$set_project("my project")
cal$add_task("first task", "jan 21")
cal$add_task("second task", "jan 23")
cal$get_workload()


## ------------------------------------------------
## Method `Calendario$get_calendar`
## ------------------------------------------------

cal <- Calendario$new()
cal$set_project("my project")
cal$add_task("first task", "jan 21")
cal$add_task("second task", "jan 23")
cal$get_workload(start = "jan 1", stop = "feb 16")
    

## ------------------------------------------------
## Method `Calendario$show_calendar`
## ------------------------------------------------

cal <- Calendario$new()
cal$set_project("my project")
cal$add_task("first task", "jan 21")
cal$add_task("second task", "jan 23")
cal$show_calendar(start = "jan 1", stop = "feb 16")
    
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Calendario-get_tasks}{\code{Calendario$get_tasks()}}
\item \href{#method-Calendario-set_project}{\code{Calendario$set_project()}}
\item \href{#method-Calendario-set_defaults}{\code{Calendario$set_defaults()}}
\item \href{#method-Calendario-get_project}{\code{Calendario$get_project()}}
\item \href{#method-Calendario-get_defaults}{\code{Calendario$get_defaults()}}
\item \href{#method-Calendario-get_options}{\code{Calendario$get_options()}}
\item \href{#method-Calendario-add_task}{\code{Calendario$add_task()}}
\item \href{#method-Calendario-show_tasks}{\code{Calendario$show_tasks()}}
\item \href{#method-Calendario-get_workload}{\code{Calendario$get_workload()}}
\item \href{#method-Calendario-get_calendar}{\code{Calendario$get_calendar()}}
\item \href{#method-Calendario-show_calendar}{\code{Calendario$show_calendar()}}
\item \href{#method-Calendario-clone}{\code{Calendario$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Calendario-get_tasks"></a>}}
\if{latex}{\out{\hypertarget{method-Calendario-get_tasks}{}}}
\subsection{Method \code{get_tasks()}}{
Retrieve a table of stored tasks
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Calendario$get_tasks(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to dplyr::filter()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A tibble with one row per task
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{cal <- Calendario$new()
cal$set_project("my project")
cal$add_task("first task", "jan 21")
cal$add_task("second task", "jan 22")
cal$get_tasks()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Calendario-set_project"></a>}}
\if{latex}{\out{\hypertarget{method-Calendario-set_project}{}}}
\subsection{Method \code{set_project()}}{
Set the default project (special case of \code{set_default()})
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Calendario$set_project(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Name of the new default project}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{cal <- Calendario$new()
cal$set_project("my project")
cal$get_project()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Calendario-set_defaults"></a>}}
\if{latex}{\out{\hypertarget{method-Calendario-set_defaults}{}}}
\subsection{Method \code{set_defaults()}}{
Set the default value for one or more specified task fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Calendario$set_defaults(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Name-value pairs specifying task fields and new default values}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{cal <- Calendario$new()
cal$set_defaults(type = "fun", team = "danielle")
cal$get_defaults()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Calendario-get_project"></a>}}
\if{latex}{\out{\hypertarget{method-Calendario-get_project}{}}}
\subsection{Method \code{get_project()}}{
Retrieve the name of the current default project
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Calendario$get_project()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector of length 1
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{cal <- Calendario$new()
cal$set_project("my project")
cal$get_project()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Calendario-get_defaults"></a>}}
\if{latex}{\out{\hypertarget{method-Calendario-get_defaults}{}}}
\subsection{Method \code{get_defaults()}}{
Retrieve the current defaults for all task fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Calendario$get_defaults()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A tibble with one row
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{cal <- Calendario$new()
cal$get_defaults()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Calendario-get_options"></a>}}
\if{latex}{\out{\hypertarget{method-Calendario-get_options}{}}}
\subsection{Method \code{get_options()}}{
Retrieve the options list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Calendario$get_options()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{cal <- Calendario$new()
cal$get_options()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Calendario-add_task"></a>}}
\if{latex}{\out{\hypertarget{method-Calendario-add_task}{}}}
\subsection{Method \code{add_task()}}{
Add a task to a project
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Calendario$add_task(
  description = NULL,
  start = NULL,
  stop = NULL,
  days = NULL,
  daily_hours = NULL,
  total_hours = NULL,
  type = NULL,
  project = NULL,
  team = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{description}}{Character string providing a description of the task}

\item{\code{start}}{Date the work starts (defaults to current date)}

\item{\code{stop}}{Date the work stops (defaults to same day as start)}

\item{\code{days}}{Number of weekdays spanned by the task}

\item{\code{daily_hours}}{Number of hours per day the task takes}

\item{\code{total_hours}}{Number of hours in total the task takes}

\item{\code{type}}{Character string assigning the task to a category}

\item{\code{project}}{Character string naming the project the task falls within}

\item{\code{team}}{Character string describing the team}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{cal <- Calendario$new()
cal$add_task(
  description = "label for the task",
  start = as.Date("2025-01-24"),
  stop = as.Date("2025-01-26"),
  daily_hours = 1,
  project = "my project",
  team = "just me"
)
cal$get_tasks()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Calendario-show_tasks"></a>}}
\if{latex}{\out{\hypertarget{method-Calendario-show_tasks}{}}}
\subsection{Method \code{show_tasks()}}{
Show a table of stored tasks as a flextable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Calendario$show_tasks(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to dplyr::filter()}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A flextable object
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{cal <- Calendario$new()
cal$set_project("my project")
cal$add_task("first task", "jan 21")
cal$add_task("second task", "jan 23")
cal$show_tasks()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Calendario-get_workload"></a>}}
\if{latex}{\out{\hypertarget{method-Calendario-get_workload}{}}}
\subsection{Method \code{get_workload()}}{
Retrieve a table describing daily workload
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Calendario$get_workload(start = NULL, stop = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start}}{Date the tabulation starts (defaults to current date)}

\item{\code{stop}}{Date the tabulation stops (defaults to start date plus 90 days)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A tibble with one row per day
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{cal <- Calendario$new()
cal$set_project("my project")
cal$add_task("first task", "jan 21")
cal$add_task("second task", "jan 23")
cal$get_workload()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Calendario-get_calendar"></a>}}
\if{latex}{\out{\hypertarget{method-Calendario-get_calendar}{}}}
\subsection{Method \code{get_calendar()}}{
Retrieve a list of tables describing monthly workload
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Calendario$get_calendar(start = NULL, stop = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start}}{Date the tabulation starts (defaults to current date)}

\item{\code{stop}}{Date the tabulation stops (defaults to start date plus 90 days)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A tibble with one row per month
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{cal <- Calendario$new()
cal$set_project("my project")
cal$add_task("first task", "jan 21")
cal$add_task("second task", "jan 23")
cal$get_workload(start = "jan 1", stop = "feb 16")
    
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Calendario-show_calendar"></a>}}
\if{latex}{\out{\hypertarget{method-Calendario-show_calendar}{}}}
\subsection{Method \code{show_calendar()}}{
Show monthly workload as a flextable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Calendario$show_calendar(start = NULL, stop = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start}}{Date the tabulation starts (defaults to current date)}

\item{\code{stop}}{Date the tabulation stops (defaults to start date plus 90 days)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A flextable object
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{cal <- Calendario$new()
cal$set_project("my project")
cal$add_task("first task", "jan 21")
cal$add_task("second task", "jan 23")
cal$show_calendar(start = "jan 1", stop = "feb 16")
    
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Calendario-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Calendario-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Calendario$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
